fix(unzip): fix Yauzl typings
Both `open` and `openReadStream` have overloads, so `promisify`
can't know which one to choose and chooses the last one (which
doesn't include options).  Additionally, once we specify those two
overloads, optional args become `arg | undefined` which is subtly
different than not providing an arg (which is what we want to be
able to do), so we specify onEntry's `openReadStream` callback
explicitly as well (to have options be optional).

I think this last point is also why we can't just update `yauzl`'s
types to have both `options` and `callback` be optional, because
`options` when `promisified` will be unioned with `undefined`, not
optional (meaning you can't just leave it off, you'd have to
actually pass `undefined` as the value if you want to use it).

https://github.com/Microsoft/TypeScript/issues/13570
https://github.com/Microsoft/TypeScript/issues/13195