fix(ios): support old simctl runtime output format
Earlier versions of `simctl` list devices by their runtime
name(`iOS 10.2`), while Xcode 10 and later list devices by their
runtime identifier (`com.apple.CoreSimulator.SimRuntime.iOS-13-0`).
There are reports that people on later versions of Xcode are
seeing the old format, so we support it just in case.

Based on those reports, although versions of Xcode 10 are using
the runtime name and not identifier, the `isAvailable` key is
still present (which is new in Xcode 10).