fix(list): handle errors with devices/virtual devices
When listing native targets, we don't want the command to fail if a
specific type can't be resolved. For example, if Xcode isn't installed
or if the Android AVD folder can't be found.

Instead of the whole command failing, just print the error to stderr and
keep the output the same, even if it means no targets can be found.

fixes https://github.com/ionic-team/native-run/issues/49